tracing

+++ rust-containers-k8s/inventory-service/src/config.rs
@@ -2,7 +2,7 @@ use serde::Deserialize;
 
 #[derive(Deserialize, Debug)]
 pub struct Config {
-    pub port: u32,
+    pub port: u16,
     pub database_url: String,
 }
